Use uppercase for `/eth/v1/config/spec` (#3325)
* Use uppercase for `/eth/v1/config/spec`

Currently, the `/eth/v1/config/spec` REST endpoint provides the response
in lowercase snake_case format. However, the specification defines this
to include the spec constants as is, without modifications. This patch
corrects this disparity using a `casingMap` applied to the `Spec` type
to force the constants to be sent in CONSTANT_CASE.

* spec casing test

Co-authored-by: harkamal <gajinder@g11.in>