Use node:* import protocol for safer module resolution (#3667)
This PR uses `node:*` imports for `fs` and `path` modules as specified in [issue 3664](https://github.com/ChainSafe/lodestar/issues/3664).

Note that using `node:*` imports requires us to whitelist `node:fs` in the `import/core-modules` `eslint` setting. If we do not whitelist it then our linter fails because it can not identify `node:fs` as a node module.