Publishes Grafana and Prometheus images for monitoring.
Prometheus only supports configuration by file, so in order to support runtime configuration via environment we have to switch over to using an entrypoint script.
`exec` will make it so `/bin/prometheus` will take over PID 1, which ensures that signals get processed correctly.
I picked some reasonable defaults for BEACON_URL and VC_URL, but I'm happy to change them to whatever.

Grafana requires building with a second context because the dashboards folder is outside of the docker folder.
This requires a relatively up-to-date version of docker and I'm *hoping* GitHub actions supports it.
Users should be able to mount a volume of their choosing targeting `/dashboards` if they want their own dashboards to be retained on top of the default set.

Grafana build was setting the admin password at build time, but that should be set at runtime.
I set the default password to `admin` under the assumption that the user will have to go out of their way to expose the service and hopefully they'll change the password when they do.  😬

I added two new build/push steps to each of the GitHub action workflows that follow the tagging convention of `chainsafe/lodestar`.
I suspect some other name will be desired like `chainsafe/lodestar-grafana` and I'm happy to change it to whatever.