Use last data point to determine nodejs version from metrics
By default Grafana sends PromQL queries to "query_range" API. Setting
"instant" to "true" means that the "query" API is used instead which is
more efficient as it only returns a single data point (end of time range).
This ensures that always the current nodejs version is shown and makes the
metric more robust in case the version changed during the selected time range
due to a Lodestar update.