Moves Prometheus entrypoint to a non-volume directory. (#4982)
Putting the entrypoint script in `/prometheus` caused problems with updating the docker image as the entrypoint was written to the volume when it was first created and then wouldn't be overwritten by future images.  Moving the entrypoint script into a non-mounted location avoids this problem.

I also added a VOLUME line to make it clear to anyone reading the dockerfile that `/prometheus` is intended to be a mounted volume, which may also help some tools auto-discover volumes and I think docker may even automatically create a volume when the image is run (not sure on the last).  Generally speaking, I believe it is a good practice to flag intended mount points in the Dockerfile.

Note: Things will work without this change, but if the entrypoint.sh file ever changes in a future image users will have a bad time when updating to the new image from an old one so ideally this should happen sooner rather than later (ideally before a lot of people start using this published image).