Print out trimmed graffiti in published block info log
The graffiti included in the block body is limited to 32 bytes which means
the length of the UTF-8 string is limited to 32 characters.

At the moment we are silently trimming the graffiti when converting it to
a buffer on the beacon node. To avoid confusion, the published block info log
printed out by the validator should display the graffiti as it was actually
included in the block body.