Add 2 seconds to default monitoring interval to avoid rate limit errors
Some calls randomly run into rate limiting, this might be due to delays when
collecting system data or other reasons which delay the previous request
which could let the next request run into a rate limit. It might also just
be due to a clock skew or other inaccuracies, it is hard to rely on time
between two remote systems. Adding 2 seconds seems to resolve this issue but
needs to be further observed.

The https://github.com/gobitfly/eth2-client-metrics-exporter also uses a
default interval of 62 seconds instead of 60 seconds.