Refactor client monitoring (#5183)
* Simplify dynamic client stats property

- remove caching as it is currently unused, system data providers handle
the caching internally

- remove support for async provider functions, not needed at the moment
as all functions are synchronous

* Add 5 seconds bucket to collect data metric

For some environments (e.g. windows) the collecting of system data might
take a bit longer. This bucket allows to get better insights on function
calls that took between 1-5 seconds.

* Add 2 seconds to default monitoring interval to avoid rate limit errors

Some calls randomly run into rate limiting, this might be due to delays when
collecting system data or other reasons which delay the previous request
which could let the next request run into a rate limit. It might also just
be due to a clock skew or other inaccuracies, it is hard to rely on time
between two remote systems. Adding 2 seconds seems to resolve this issue but
needs to be further observed.

The https://github.com/gobitfly/eth2-client-metrics-exporter also uses a
default interval of 62 seconds instead of 60 seconds.

* Add comment on why only remote host should be logged

* Override colors in send data duration panel

Use color red for errors and color green for successful requests to remote service.

* Add jsdoc to client stats properties

* Use CONSTANT_CASE for client stats schema fixtures