chore: fix spec and e2e tests for peerDAS (#7713)
**Motivation**

Fixes an error in the spec test, as well as an error in the e2e tests.
This might not fix all the e2e test issues, but I still need to get my
e2e environment set up locally.

**Description**

* Continues skipping the test for single_merkle_proof. This is removed
for fulu in 1.5.0-beta4 tests, but might as well fix it for now
* Allows the params e2e tests to run on Fulu -- everything is
implemented there anyway